+++ src/resources/views/partials/model_table_actions.blade.php
@@ -1,2 +1,6 @@
-<a href="{{ $crudderModel->editUrl($model) }}">Edit</a>
-<a href="{{ $crudderModel->deleteUrl($model) }}">Delete</a>
+@if($crudderModel->actionEnabled('update'))
+    <a href="{{ $crudderModel->editUrl($model) }}">Edit</a>
+@endif
+@if($crudderModel->actionEnabled('delete'))
+    <a href="{{ $crudderModel->deleteUrl($model) }}">Delete</a>
+@endif
